pyproject.toml
[tool.poetry]
name = "repl_python3_bitly"
version = "0.1.0"
description = ""
authors = ["dwreaper <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
